switch size for storage keys to size_t ... much easier to use strlen() and sizeof() directly without casts

+++ examples/chip-tool/config/PersistentStorage.h
@@ -29,7 +29,7 @@ public:
     chip::Logging::LogCategory GetLoggingLevel();
 
     /////////// PersistentStorageDelegate Interface /////////
-    CHIP_ERROR SyncGetKeyValue(const char * key, void * value, uint16_t & size) override;
-    CHIP_ERROR SyncSetKeyValue(const char * key, const void * value, uint16_t size) override;
+    CHIP_ERROR SyncGetKeyValue(const char * key, void * value, size_t & size) override;
+    CHIP_ERROR SyncSetKeyValue(const char * key, const void * value, size_t size) override;
     CHIP_ERROR SyncDeleteKeyValue(const char * key) override;
 };
