+++ src/controller/java/AndroidDeviceControllerWrapper.cpp
@@ -285,7 +285,7 @@ void AndroidDeviceControllerWrapper::OnMessage(chip::System::PacketBufferHandle 
 
 void AndroidDeviceControllerWrapper::OnStatusChange(void) {}
 
-CHIP_ERROR AndroidDeviceControllerWrapper::SyncGetKeyValue(const char * key, char * value, uint16_t & size)
+CHIP_ERROR AndroidDeviceControllerWrapper::SyncGetKeyValue(const char * key, char * value, size_t & size)
 {
     jstring keyString       = NULL;
     jstring valueString     = NULL;
