+++ src/controller/python/chip/internal/CommissionerImpl.cpp
@@ -28,12 +28,12 @@ namespace {
 class ServerStorageDelegate : public chip::PersistentStorageDelegate
 {
 public:
-    CHIP_ERROR SyncSetKeyValue(const char * key, const void * value, uint16_t size) override
+    CHIP_ERROR SyncSetKeyValue(const char * key, const void * value, size_t size) override
     {
         return chip::DeviceLayer::PersistedStorage::KeyValueStoreMgr().Put(key, value, size);
     }
 
-    CHIP_ERROR SyncGetKeyValue(const char * key, void * value, uint16_t & size) override
+    CHIP_ERROR SyncGetKeyValue(const char * key, void * value, size_t & size) override
     {
         return chip::DeviceLayer::PersistedStorage::KeyValueStoreMgr().Get(key, value, size);
     }
