+++ src/darwin/Framework/CHIP/CHIPPersistentStorageDelegateBridge.mm
@@ -26,7 +26,7 @@ CHIPPersistentStorageDelegateBridge::CHIPPersistentStorageDelegateBridge(void)
 
 CHIPPersistentStorageDelegateBridge::~CHIPPersistentStorageDelegateBridge(void) {}
 
-CHIP_ERROR CHIPPersistentStorageDelegateBridge::SyncGetKeyValue(const char * key, void * value, uint16_t & size)
+CHIP_ERROR CHIPPersistentStorageDelegateBridge::SyncGetKeyValue(const char * key, void * value, size_t & size)
 {
     /*
         TODO: read from a suitable binary encoding
@@ -69,7 +69,7 @@ CHIP_ERROR CHIPPersistentStorageDelegateBridge::SyncGetKeyValue(const char * key
     return CHIP_ERROR_NOT_IMPLEMENTED;
 }
 
-CHIP_ERROR CHIPPersistentStorageDelegateBridge::SyncSetKeyValue(const char * key, const void * value, uint16_t size)
+CHIP_ERROR CHIPPersistentStorageDelegateBridge::SyncSetKeyValue(const char * key, const void * value, size_t size)
 {
     /*
         TODO: convert value to some suitable binary encoding and store
